Rename result-table to resolve-table
