+++ components/index.js
@@ -1,8 +1,8 @@
 export * from './search-box/search-box'
 export * from './options-box/options-box'
-export * from './result-table/result-table'
+export * from './resolve-table/resolve-table'
 export * from './footer/footer'
 export * from './top-bar/top-bar'
 export * from './download-results/download-results'
 export * from './parse-table/parse-table'
-export * from './result-table/best-match-settings-popper'
+export * from './resolve-table/best-match-settings-popper'
