+++ components/resolve-table/details-dialog.jsx
