+++ components/resolve-table/links.jsx
