+++ components/resolve-table/select-row.jsx
