+++ components/resolve-table/warnings.jsx
