--- components/result-table/result-table.style.js
@@ -1,4 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles'
-
-export const useStyles = makeStyles(() => ({
-}))
